[feat] 아티스트 API 완료

See merge request s13-webmobile1-sub1/S13P11A404!23

+++ backend/BE/src/main/java/com/a404/duckonback/entity/ArtistFollowId.java
@@ -9,5 +9,5 @@ import java.io.Serializable;
 @EqualsAndHashCode
 public class ArtistFollowId implements Serializable {
     private Long user;       // User.id
-    private Integer artist;    // Artist.artistId
+    private Long artist;    // Artist.artistId
 }
